feat: Move Nodejs4.3/6.10 to a legacy template, add sam build under bringing to the next level
