+++ samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs-legacy/cookiecutter.json
@@ -1,0 +1,4 @@
+{
+    "project_name": "Name of the project",
+    "runtime": "nodejs8.10"
+}
