+++ samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs-legacy/setup.cfg
@@ -1,0 +1,2 @@
+[install]
+prefix=
