+++ samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs-legacy/{{cookiecutter.project_name}}/hello-world/.npmignore
@@ -1,0 +1,1 @@
+tests/*
