+++ samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs/{{cookiecutter.project_name}}/hello-world/app.js
@@ -1,3 +1,7 @@
+// const axios = require('axios')
+// const url = 'http://checkip.amazonaws.com/';
+let response;
+
 /**
  *
  * Event doc: https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-input-format
@@ -10,26 +14,20 @@
  * @returns {Object} object - API Gateway Lambda Proxy Output Format
  * 
  */
-{%- if cookiecutter.runtime == 'nodejs6.10' or cookiecutter.runtime == 'nodejs4.3' %}
-exports.lambdaHandler = function (event, context, callback) {
-    var response = {
-        statusCode: 200,
-        body: JSON.stringify({
-            message: 'hello world'
-        })
-    }
-
-    callback(null, response);
-};
-{%- else %}
 exports.lambdaHandler = async (event, context) => {
-    var response = {
-        statusCode: 200,
-        body: JSON.stringify({
-            message: 'hello world'
-        })
+    try {
+        // const ret = await axios(url);
+        response = {
+            'statusCode': 200,
+            'body': JSON.stringify({
+                message: 'hello world',
+                // location: ret.data.trim()
+            })
+        }
+    } catch (err) {
+        console.log(err);
+        return err;
     }
 
     return response
 };
-{% endif %}
