+++ samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs/{{cookiecutter.project_name}}/hello-world/tests/unit/test-handler.js
@@ -5,28 +5,6 @@ const chai = require('chai');
 const expect = chai.expect;
 var event, context;
 
-{% if cookiecutter.runtime == 'nodejs6.10' or cookiecutter.runtime == 'nodejs4.3' %}
-describe('Tests Handler', function () {
-    it('verifies successful response', function (done) {
-        app.lambdaHandler(event, context, function (err, result) {
-            try {
-                expect(result).to.be.an('object');
-                expect(result.statusCode).to.equal(200);
-                expect(result.body).to.be.an('string');
-
-                let response = JSON.parse(result.body);
-
-                expect(response).to.be.an('object');
-                expect(response.message).to.be.equal("hello world");
-                expect(response.location).to.be.an("string");
-                done();
-            } catch (e) {
-                done(e);
-            }
-        });
-    });
-});
-{% else %}
 describe('Tests index', function () {
     it('verifies successful response', async () => {
         const result = await app.lambdaHandler(event, context)
@@ -39,7 +17,6 @@ describe('Tests index', function () {
 
         expect(response).to.be.an('object');
         expect(response.message).to.be.equal("hello world");
-        expect(response.location).to.be.an("string");
+        // expect(response.location).to.be.an("string");
     });
 });
-{% endif %}
