--- samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs/{{cookiecutter.project_name}}/hello_world/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "hello_world",
-  "version": "1.0.0",
-  "description": "hello world sample for NodeJS",
-  "author": "SAM CLI",
-  "license": "MIT",
-  "scripts": {
-    "test": "mocha tests/unit/",
-    "start": "cd .. && sam local start-api",
-    "invoke": "cd .. && sam local invoke HelloWorldFunction -e event.json"
-  },
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "mocha": "^5.1.1"
-  }
-}
